BlobStore: use normal func ptr instead of member func ptr
see https://gcc.gnu.org/onlinedocs/gcc/Bound-member-functions.html

normal func ptr is more efficient, and the calling conversion is
compatible to member func ptr, the extra is `this` pointer ajustion
and virtual function.

member function used in BlobStore is compatible to normal func ptr,
it need not `this` pointer adjustion, and no virtual function.

So we use normal function ptr instead of member func ptr to improve
performance.

msvc member func ptr is not wide pointer, it is more efficient than
gcc, so it is not changed in msvc.